Create version1
Library for running simulations and displaying data. Inputs are the dimensions of the simulations, number of simulations, data plotting and display inputs. Adds simulation data to a file containing data for simulations with the given dimensions, and plots the first particle trajectory, the initial coordinates of all simulations in the file, velocity moduli, the average velocities wrt time, and the frequency of consecutive bounces, being 0, 1, 2, 3, 4, 5, 6, 7, 8<=